pyproject.toml
[tool.poetry]
name = "brscans-backend"
version = "0.1.0"
description = ""
authors = ["Samuel Pereira da Silva <samuel.silva513@etec.sp.gov.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "4.2.9"
djangorestframework = "3.14.0"
black = "23.12.1"
httpx = "0.26.0"
bs4 = "0.0.2"
unidecode = "1.3.8"
requests = "^2.32.3"
django-cors-headers = "^4.4.0"
whitenoise = "^6.7.0"
gunicorn = "^22.0.0"
cloudscraper = "^1.2.71"
drf-nested-routers = "^0.94.1"
pillow = "^10.4.0"
xmltodict = "^0.13.0"
psycopg2-binary = "^2.9.9"
django-storages = "^1.14.3"
zappa = "^0.59.0"
tldextract = "^5.1.2"
replicate = "^0.28.0"
django-debug-toolbar = "^4.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
